% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseComplementAlternateCase.R
\name{reverseComplementAlternateCase}
\alias{reverseComplementAlternateCase}
\title{Reverse Complement of a DNA Sequence with Alternating Case
E.g.: "ATCG" -> "CgAt"}
\usage{
reverseComplementAlternateCase(s)
}
\arguments{
\item{s}{A string representing a DNA sequence.}
}
\value{
A string representing the reverse complement of the input DNA sequence, with alternating case.
}
\description{
Reverse Complement of a DNA Sequence with Alternating Case
E.g.: "ATCG" -> "CgAt"
}
